Restore PR #448 with pick of #447 for `tty_signal_keys`
* Vterm is fully not supported under windows
* Do not import ZMQEventLoop on windows due to python crash
* rework test for event loops to be windows compatible: on windows `select` do not support pipe
* on windows11 console encoding is detected as cp1252, BTW it's validated manually that cms uses UTF-8
* fix 24 bit colors terminal

Skip tests:
* ZMQEventLoop is not supported under windows
* `AsyncioEventLoop`, `TornadoEventLoop` and `TwistedEventLoop`:
  * Issues with asyncio loop (changes between python version, select support)
  * Twisted and tornado tests show race conditions
