+++ tests/test_doctests.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import doctest
+import os
 import unittest
 
 import urwid
@@ -43,6 +44,13 @@ def load_tests(loader: unittest.TestLoader, tests: unittest.BaseTestSuite, ignor
         urwid.util,
         urwid.signals,
     ]
+
+    option_flags = doctest.ELLIPSIS | doctest.DONT_ACCEPT_TRUE_FOR_1 | doctest.IGNORE_EXCEPTION_DETAIL
     for m in module_doctests:
-        tests.addTests(doctest.DocTestSuite(m, optionflags=doctest.ELLIPSIS | doctest.IGNORE_EXCEPTION_DETAIL))
+        tests.addTests(doctest.DocTestSuite(m, optionflags=option_flags))
+
+    if os.name == "nt":
+        tests.addTests(doctest.DocTestSuite("urwid._win32_raw_display", optionflags=option_flags))
+    else:
+        tests.addTests(doctest.DocTestSuite("urwid._posix_raw_display", optionflags=option_flags))
     return tests
