+++ tests/test_event_loops.py
@@ -2,31 +2,63 @@ from __future__ import annotations
 
 import asyncio
 import os
+import socket
 import sys
 import typing
 import unittest
 
 import urwid
 
+if typing.TYPE_CHECKING:
+    from types import TracebackType
+
+IS_WINDOWS = os.name == "nt"
+
+
+class ClosingSocketPair(typing.ContextManager[typing.Tuple[socket.socket, socket.socket]]):
+    __slots__ = ("rd_s", "wr_s")
+
+    def __init__(self) -> None:
+        self.rd_s: socket.socket | None = None
+        self.wr_s: socket.socket | None = None
+
+    def __enter__(self) -> tuple[socket.socket, socket.socket]:
+        self.rd_s, self.wr_s = socket.socketpair()
+        return self.rd_s, self.wr_s
+
+    def __exit__(
+        self,
+        exc_type: type[BaseException] | None,
+        exc_val: BaseException | None,
+        exc_tb: TracebackType | None,
+    ) -> None:
+        if self.rd_s is not None:
+            self.rd_s.close()
+        if self.wr_s is not None:
+            self.wr_s.close()
+
 
 class EventLoopTestMixin:
     def test_event_loop(self):
-        rd, wr = os.pipe()
         evl: urwid.EventLoop = self.evl
         out = []
+        rd: socket.socket
+        wr: socket.socket
 
         def step1():
             out.append("writing")
-            os.write(wr, b"hi")
+            wr.send(b"hi")
 
         def step2():
-            out.append(os.read(rd, 2).decode('ascii'))
+            out.append(rd.recv(2).decode("ascii"))
             raise urwid.ExitMainLoop
 
-        _handle = evl.alarm(0, step1)
-        _handle = evl.watch_file(rd, step2)
+        with ClosingSocketPair() as (rd, wr):
+            _handle = evl.alarm(0, step1)
+            _handle = evl.watch_file(rd.fileno(), step2)
 
-        evl.run()
+            evl.run()
+
         self.assertEqual(out, ["writing", "hi"])
 
     def test_remove_alarm(self):
@@ -43,10 +75,9 @@ class EventLoopTestMixin:
 
     def test_remove_watch_file(self):
         evl: urwid.EventLoop = self.evl
-        fd_r, fd_w = os.pipe()
 
-        try:
-            handle = evl.watch_file(fd_r, lambda: None)
+        with ClosingSocketPair() as (rd, _wr):
+            handle = evl.watch_file(rd.fileno(), lambda: None)
 
             def step1():
                 self.assertTrue(evl.remove_watch_file(handle))
@@ -55,17 +86,14 @@ class EventLoopTestMixin:
 
             evl.alarm(0, step1)
             evl.run()
-        finally:
-            os.close(fd_r)
-            os.close(fd_w)
 
     _expected_idle_handle = 1
 
     def test_run(self):
         evl: urwid.EventLoop = self.evl
         out = []
-        rd, wr = os.pipe()
-        self.assertEqual(os.write(wr, b"data"), 4)
+        wr: socket.socket
+        rd: socket.socket
 
         def say_hello():
             out.append("hello")
@@ -78,28 +106,31 @@ class EventLoopTestMixin:
             raise urwid.ExitMainLoop
 
         def exit_error() -> typing.NoReturn:
-            1/0
+            1 / 0
 
-        _handle = evl.alarm(0.01, exit_clean)
-        _handle = evl.alarm(0.005, say_hello)
-        idle_handle = evl.enter_idle(say_waiting)
-        if self._expected_idle_handle is not None:
-            self.assertEqual(idle_handle, 1)
+        with ClosingSocketPair() as (rd, wr):
+            self.assertEqual(wr.send(b"data"), 4)
 
-        evl.run()
-        self.assertTrue("waiting" in out, out)
-        self.assertTrue("hello" in out, out)
-        self.assertTrue("clean exit" in out, out)
-        handle = evl.watch_file(rd, exit_clean)
-        del out[:]
+            _handle = evl.alarm(0.01, exit_clean)
+            _handle = evl.alarm(0.005, say_hello)
+            idle_handle = evl.enter_idle(say_waiting)
+            if self._expected_idle_handle is not None:
+                self.assertEqual(idle_handle, 1)
 
-        evl.run()
-        self.assertEqual(["clean exit"], out)
-        self.assertTrue(evl.remove_watch_file(handle))
-        _handle = evl.alarm(0, exit_error)
-        self.assertRaises(ZeroDivisionError, evl.run)
-        _handle = evl.watch_file(rd, exit_error)
-        self.assertRaises(ZeroDivisionError, evl.run)
+            evl.run()
+            self.assertTrue("waiting" in out, out)
+            self.assertTrue("hello" in out, out)
+            self.assertTrue("clean exit" in out, out)
+            handle = evl.watch_file(rd.fileno(), exit_clean)
+            del out[:]
+
+            evl.run()
+            self.assertEqual(["clean exit"], out)
+            self.assertTrue(evl.remove_watch_file(handle))
+            _handle = evl.alarm(0, exit_error)
+            self.assertRaises(ZeroDivisionError, evl.run)
+            _handle = evl.watch_file(rd.fileno(), exit_error)
+            self.assertRaises(ZeroDivisionError, evl.run)
 
 
 class SelectEventLoopTest(unittest.TestCase, EventLoopTestMixin):
@@ -112,6 +143,7 @@ try:
 except ImportError:
     pass
 else:
+
     class GLibEventLoopTest(unittest.TestCase, EventLoopTestMixin):
         def setUp(self):
             self.evl = urwid.GLibEventLoop()
@@ -127,9 +159,12 @@ try:
 except ImportError:
     pass
 else:
+
+    @unittest.skipIf(IS_WINDOWS, "Windows is temporary not supported by TornadoEventLoop.")
     class TornadoEventLoopTest(unittest.TestCase, EventLoopTestMixin):
         def setUp(self):
             from tornado.ioloop import IOLoop
+
             self.evl = urwid.TornadoEventLoop(IOLoop())
 
         _expected_idle_handle = None
@@ -140,6 +175,8 @@ try:
 except ImportError:
     pass
 else:
+
+    @unittest.skipIf(IS_WINDOWS, "Windows is temporary not supported by TwistedEventLoop.")
     class TwistedEventLoopTest(unittest.TestCase, EventLoopTestMixin):
         def setUp(self):
             self.evl = urwid.TwistedEventLoop()
@@ -157,11 +194,11 @@ else:
         def test_run(self):
             evl = self.evl
             out = []
-            rd, wr = os.pipe()
-            self.assertEqual(os.write(wr, b"data"), 4)
+            wr: socket.socket
+            rd: socket.socket
 
             def step2():
-                out.append(os.read(rd, 2).decode('ascii'))
+                out.append(rd.recv(2).decode("ascii"))
 
             def say_hello():
                 out.append("hello")
@@ -176,14 +213,11 @@ else:
                 out.append("remove_alarm ok")
 
             def test_remove_watch_file():
-                fd_r, fd_w = os.pipe()
-                try:
-                    handle = evl.watch_file(fd_r, lambda: None)
+                with ClosingSocketPair() as (rd_, _wr):
+                    handle = evl.watch_file(rd_.fileno(), lambda: None)
                     self.assertTrue(evl.remove_watch_file(handle))
                     self.assertFalse(evl.remove_watch_file(handle))
-                finally:
-                    os.close(fd_r)
-                    os.close(fd_w)
+
                 out.append("remove_watch_file ok")
 
             def exit_clean() -> typing.NoReturn:
@@ -191,15 +225,19 @@ else:
                 raise urwid.ExitMainLoop
 
             def exit_error() -> typing.NoReturn:
-                1/0
+                1 / 0
 
-            _handle = evl.watch_file(rd, step2)
-            _handle = evl.alarm(0.1, exit_clean)
-            _handle = evl.alarm(0.05, say_hello)
-            _handle = evl.alarm(0.06, test_remove_alarm)
-            _handle = evl.alarm(0.07, test_remove_watch_file)
-            self.assertEqual(evl.enter_idle(say_waiting), 1)
-            evl.run()
+            with ClosingSocketPair() as (rd, wr):
+                self.assertEqual(wr.send(b"data"), 4)
+
+                _handle = evl.watch_file(rd.fileno(), step2)
+                _handle = evl.alarm(0.1, exit_clean)
+                _handle = evl.alarm(0.05, say_hello)
+                _handle = evl.alarm(0.06, test_remove_alarm)
+                _handle = evl.alarm(0.07, test_remove_watch_file)
+                self.assertEqual(evl.enter_idle(say_waiting), 1)
+                evl.run()
+
             self.assertTrue("da" in out, out)
             self.assertTrue("ta" in out, out)
             self.assertTrue("hello" in out, out)
@@ -212,6 +250,7 @@ else:
             self.assertRaises(ZeroDivisionError, evl.run)
 
 
+@unittest.skipIf(IS_WINDOWS, "Windows is temporary not supported by AsyncioEventLoop.")
 class AsyncioEventLoopTest(unittest.TestCase, EventLoopTestMixin):
     def setUp(self):
         self.evl = urwid.AsyncioEventLoop()
@@ -243,6 +282,7 @@ try:
 except ImportError:
     pass
 else:
+
     class TrioEventLoopTest(unittest.TestCase, EventLoopTestMixin):
         def setUp(self):
             self.evl = urwid.TrioEventLoop()
@@ -260,6 +300,8 @@ try:
 except ImportError:
     pass
 else:
+
+    @unittest.skipIf(IS_WINDOWS, "ZMQEventLoop is not supported under windows")
     class ZMQEventLoopTest(unittest.TestCase, EventLoopTestMixin):
         def setUp(self):
             self.evl = urwid.ZMQEventLoop()
