+++ urwid/__init__.py
@@ -21,6 +21,8 @@
 
 from __future__ import annotations
 
+import os
+
 from urwid import raw_display
 from urwid.canvas import (
     BlankCanvas,
@@ -121,7 +123,6 @@ from urwid.util import (
 )
 from urwid.version import version as __version__
 from urwid.version import version_tuple as __version_tuple__
-from urwid.vterm import TermCanvas, TermCharset, Terminal, TermModes, TermScroller
 from urwid.widget import (
     ANY,
     BOTTOM,
@@ -229,10 +230,15 @@ try:
 except ImportError:
     pass
 
-try:
-    from urwid.event_loop import ZMQEventLoop
-except ImportError:
-    pass
+# OS Specific
+if os.name != "nt":
+    from urwid.vterm import TermCanvas, TermCharset, Terminal, TermModes, TermScroller
+
+    # ZMQEventLoop cause interpreter crash on windows
+    try:
+        from urwid.event_loop import ZMQEventLoop
+    except ImportError:
+        pass
 
 # Backward compatibility
 VERSION = __version_tuple__
